Add files via upload
Task 41: Evaluate various combinations of logical operators (&&, ||, !). 
Task 42: Use logical operators to write a condition that checks if a number is in a 
given range. 
Task 43: Use the NOT (!) operator to invert a boolean value. 
Task 44: Evaluate the short-circuiting nature of logical operators. 
Task 45:Compare two non-boolean values using logical operators and observe 
the result.